Pictures Added

+++ src/Gui/Logs/jViewLogs.java
@@ -6,6 +6,7 @@ import Gui.jMainPage;
 import Services.LogServices;
 import java.util.ArrayList;
 import java.util.UUID;
+import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
 import utils.FilterLogs;
 
@@ -62,12 +63,15 @@ public class jViewLogs extends javax.swing.JPanel {
     public void delete(){
         DefaultTableModel m = (DefaultTableModel) jLogsTable.getModel();
         if(jLogsTable.getSelectedRow() != -1) {
-        UUID id = (UUID) m.getValueAt(jLogsTable.getSelectedRow(), 0);
-        m.removeRow(jLogsTable.getSelectedRow());
-        _LogServices.deleteLog(id);
-        jDeleteMessage.setText("Deleted Successfully");
+            UUID id = (UUID) m.getValueAt(jLogsTable.getSelectedRow(), 0);
+            int answer = JOptionPane.showConfirmDialog(null, "Are you Sure you want to delete this Log?");
+            if(answer == JOptionPane.YES_OPTION) {
+                m.removeRow(jLogsTable.getSelectedRow());
+                _LogServices.deleteLog(id);
+                JOptionPane.showMessageDialog(null, "Deleted Successfully!");
+            }
         }else{
-            jDeleteMessage.setText("Error! Please Select Log");
+            JOptionPane.showMessageDialog(null, "Please Select a Log to be Deleted!");
         }
         renderData();
     }
@@ -99,7 +103,6 @@ public class jViewLogs extends javax.swing.JPanel {
         jLabel3 = new javax.swing.JLabel();
         jScrollPane1 = new javax.swing.JScrollPane();
         jLogDetails = new javax.swing.JTextArea();
-        jDeleteMessage = new javax.swing.JLabel();
         jScrollPane3 = new javax.swing.JScrollPane();
         jLogsTable = new javax.swing.JTable();
 
@@ -176,9 +179,6 @@ public class jViewLogs extends javax.swing.JPanel {
         jLogDetails.setRows(5);
         jScrollPane1.setViewportView(jLogDetails);
 
-        jDeleteMessage.setFont(new java.awt.Font("Calibri", 0, 18)); // NOI18N
-        jDeleteMessage.setForeground(new java.awt.Color(0, 31, 78));
-
         jLogsTable.setBackground(new java.awt.Color(217, 156, 69));
         jLogsTable.setFont(new java.awt.Font("Calibri", 0, 18)); // NOI18N
         jLogsTable.setForeground(new java.awt.Color(255, 255, 255));
@@ -242,22 +242,20 @@ public class jViewLogs extends javax.swing.JPanel {
                             .addComponent(jScrollPane3)
                             .addComponent(jScrollPane1)
                             .addGroup(layout.createSequentialGroup()
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                    .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                    .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                                .addGap(18, 18, 18)
                                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(jSearchField, javax.swing.GroupLayout.PREFERRED_SIZE, 506, javax.swing.GroupLayout.PREFERRED_SIZE)
                                     .addGroup(layout.createSequentialGroup()
-                                        .addComponent(jSortByCombo, javax.swing.GroupLayout.PREFERRED_SIZE, 506, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                            .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                            .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
                                         .addGap(18, 18, 18)
-                                        .addComponent(jToggleSort)))
-                                .addGap(0, 20, Short.MAX_VALUE))
-                            .addGroup(layout.createSequentialGroup()
-                                .addComponent(jLabel3)
-                                .addGap(18, 18, 18)
-                                .addComponent(jDeleteMessage, javax.swing.GroupLayout.PREFERRED_SIZE, 324, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                            .addComponent(jSearchField, javax.swing.GroupLayout.PREFERRED_SIZE, 506, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                            .addGroup(layout.createSequentialGroup()
+                                                .addComponent(jSortByCombo, javax.swing.GroupLayout.PREFERRED_SIZE, 506, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(18, 18, 18)
+                                                .addComponent(jToggleSort))))
+                                    .addComponent(jLabel3))
+                                .addGap(0, 20, Short.MAX_VALUE)))
                         .addGap(30, 30, 30))))
         );
         layout.setVerticalGroup(
@@ -274,9 +272,7 @@ public class jViewLogs extends javax.swing.JPanel {
                         .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                         .addComponent(jToggleSort)))
                 .addGap(26, 26, 26)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel3)
-                    .addComponent(jDeleteMessage))
+                .addComponent(jLabel3)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 141, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(18, 18, 18)
@@ -292,7 +288,6 @@ public class jViewLogs extends javax.swing.JPanel {
     private void jBackMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jBackMouseClicked
         _jHomePage.switchPanels(_jMainPage);
         _jMainPage.jLog.grabFocus();
-        jDeleteMessage.setText("");
     }//GEN-LAST:event_jBackMouseClicked
 
     private void jDeleteMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jDeleteMouseClicked
@@ -302,7 +297,6 @@ public class jViewLogs extends javax.swing.JPanel {
     private void jToggleSortMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jToggleSortMouseClicked
         toggle = !toggle;
         updateData();
-        jDeleteMessage.setText("");
     }//GEN-LAST:event_jToggleSortMouseClicked
 
     private void jSearchFieldKeyTyped(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jSearchFieldKeyTyped
@@ -311,11 +305,11 @@ public class jViewLogs extends javax.swing.JPanel {
 
 
     private void jSearchFieldMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jSearchFieldMouseClicked
-        jDeleteMessage.setText("");
+        
     }//GEN-LAST:event_jSearchFieldMouseClicked
 
     private void jSortByComboMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jSortByComboMouseClicked
-        jDeleteMessage.setText("");
+        
     }//GEN-LAST:event_jSortByComboMouseClicked
 
     private void jLogsTableMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLogsTableMousePressed
@@ -332,7 +326,6 @@ public class jViewLogs extends javax.swing.JPanel {
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton jBack;
     private javax.swing.JButton jDelete;
-    private javax.swing.JLabel jDeleteMessage;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
